backend/requirements.txt
absl-py==0.9.0
appdirs==1.4.4
asgiref==3.2.10
astunparse==1.6.3
awscli==1.18.152
backcall==0.2.0
beautifulsoup4==4.9.2
boto3==1.15.11
botocore==1.18.11
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.2
django-allauth==0.42.0
django-cors-headers==3.5.0
django-extensions==3.0.9
django-rest-auth==0.9.5
django-storages==1.10.1
djangorestframework==3.11.1
docutils==0.15.2
filelock==3.0.12
future==0.18.2
gast==0.3.3
geojson==2.5.0
google-auth==1.20.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
idna==2.10
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
jmespath==0.10.0
jwt==1.0.0
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.3.2
numpy==1.19.2
oauthlib==3.1.0
opencv-python==4.4.0.44
opt-einsum==3.3.0
optional-django==0.3.0
parso==0.7.1
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.7
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.1
pyowm==3.1.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
react==4.3.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.5
s3transfer==0.3.3
scipy==1.4.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
torch==1.6.0
torchvision==0.7.0
tqdm==4.50.0
traitlets==5.0.4
urllib3==1.25.10
virtualenv==20.0.31
wcwidth==0.2.5
Werkzeug==1.0.1
